Fix the current Windows build
As logged in `bld.bat`: as of July 2023, the build system for this
package can run into issues if the version of `cl` found via %PATH% is
not exactly the same as the version used by MSBuild. The only way I can
figure out to avoid the problem is to avoid MSBuild altogether, which
can be done by using the NMake Makefiles generator, as well as
explicitly specifying the full path to the desired compilers.

The error messages were:

```
Exception in file D:/dev/cppan2/client2/src/sw/driver/module.cpp:120, function operator ():
  error in module (C:/Users/VssAdministrator/.sw/storage/tmp/cfg/793929/loc.sw.self.4a156b-0.0.31.dll):
    Exception in file D:/dev/cppan2/client2/src/sw/driver/target/native.cpp:799, function findCompiler:
      Cannot find compiler com.Microsoft.VisualStudio.VC.cl->=19.35.0 <19.36.0 for settings: {...}
```