Don't use `sw` when building on Windows
After all that effort to get it to work ... it turns out that the build just
works on Windows if we don't use `sw`, and just pick up the conda-forge
dependencies. And in my testing, we need a non-`sw` build in order to
successfully build Tesseract against the newest Leptonica.